# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a API-Jira server.",
        "version": "0.0.1",
        "title": "Jira-API",
        "termsOfService": ""
    },
    "host": "",
    "basePath": "/",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/get/jira/object/objectSchemaNametoID": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "HGC_API_TEST"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/objectSchemaKeytoID": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "HGCAPI"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/objectTypeNametoID": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "14"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "DC-Customer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/objectNametoID": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "14"
                    },
                    {
                        "name": "objectType",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "DC-Customer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/objects": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "16"
                    },
                    {
                        "name": "objectTypeId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "202"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/object": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectId",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "255694"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/objectsWithNames": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Asset Management"
                    },
                    {
                        "name": "objectTypeName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "CPE Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/objectsWithNamesAttributes": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Asset Management"
                    },
                    {
                        "name": "objectTypeName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "CPE Type"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/get/jira/object/attributeValue": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "TOC"
                    },
                    {
                        "name": "objectTypeName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Assignment User"
                    },
                    {
                        "name": "findAttribute",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Group"
                    },
                    {
                        "name": "findValue",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "TOC"
                    },
                    {
                        "name": "returnAttribute",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Email"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/get/jira/object/includeAttributObject": {
            "get": {
                "tags": [
                    "Get Jira Objects"
                ],
                "summary": "",
                "description": "",
                "operationId": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "objectSchemaName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "IBVSD"
                    },
                    {
                        "name": "objectTypeName",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Customer"
                    },
                    {
                        "name": "attribute",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "Sender"
                    },
                    {
                        "name": "value",
                        "in": "query",
                        "required": true,
                        "type": "string",
                        "default": "herbert.tung"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}